refactor: fix some styles. remove unnessary code. rename some variables
