+++ frontend/src/app/app.scss
@@ -1,0 +1,7 @@
+.main {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  flex: 1;
+}
