--- frontend/src/app/core/components/test-button/test-button.html
@@ -1,1 +1,0 @@
-<button (click)="simulateNetworkError()">Спровоцировать ошибку</button>
