+++ frontend/src/app/core/layout/header/header.html
@@ -4,7 +4,6 @@
       <img class="logo_image" src="assets/images/logo.svg" alt="logo" />
     </a>
     <div class="services">
-      <app-test-button></app-test-button>
       <app-theme-switcher></app-theme-switcher>
       <app-language-switcher></app-language-switcher>
     </div>
