+++ frontend/src/app/core/layout/header/header.ts
@@ -4,11 +4,10 @@ import { ThemeSwitcher } from '../../components/theme-switcher/theme-switcher';
 import { TuiAppearance } from '@taiga-ui/core';
 import { RouterLink } from '@angular/router';
 import { LaguageSwitcher } from '../../components/language-switcher/language-switcher';
-import { TestButton } from '../../components/test-button/test-button';
 
 @Component({
   selector: 'app-header',
-  imports: [TuiHeader, RouterLink, TuiAppearance, LaguageSwitcher, ThemeSwitcher, TestButton],
+  imports: [TuiHeader, RouterLink, TuiAppearance, LaguageSwitcher, ThemeSwitcher],
   templateUrl: './header.html',
   styleUrl: './header.scss',
 })
