--- frontend/src/app/shared/constants/enum.ts
@@ -1,16 +1,0 @@
-export enum ThemeNames {
-  Light = 'light',
-  Dark = 'dark',
-}
-
-export enum ErrorTosterLabels {
-  Error = 'error',
-  Warning = 'warning',
-  Positive = 'meow',
-}
-
-export enum ErrorTosterAppearances {
-  Error = 'negative',
-  Warning = 'warning',
-  Positive = 'positive',
-}
