+++ frontend/src/app/shared/enums/theme-names.enum.ts
@@ -1,0 +1,4 @@
+export enum ThemeNames {
+  Light = 'light',
+  Dark = 'dark',
+}
